RTT sorunu halledildi. Server side da iki class oluşturuldu. operation class'ı artık serverin operasyonun yapılacağı yerdir. serverAnalysis class'ı server daki raporlama kısmının yapıldığı kısımdır.
malik
+++Clientleri tek tek runlayınca ilk sensorün rtt hep 0, diğerlerinin hep 1 oluyor. - düzeldi yaptım

ClientSide
+++ clientside da iki yeni class oluşturuldu.
randomVariable : bu bizim random variable generate ettiğimiz class. daha öncesinde sendObjects clasının içerisinde
bir method yardımı ile yapıyorduk.
ClientAnalysis : bu bizim analiz ve raporlama classımız. daha öncesinde sendobject classımız içerisinde yapıyorduk.

ServerSide
serverside da artık iki yeni classımız var bunlar.
ServerOperation : bu classımız iş yükünü yapan class ımız. daha öncesinde printQueue classında bir methoddu.
ServerAnalysis : bu classımız sayesinde artık serverdaki raporlamayı yapıyoruz. printQueue deydi buda methodlar olarak.
